pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Basic image anomaly detection"
readme = "README.md"
license = {text = "MIT"}
authors = [{ name="Sara Hatami", email="sarahatamig@gmail.com"}]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "scikit-image>=0.22",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "reportlab>=4.0",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "coverage>=7.4",
    "black>=24.3",
    "ruff>=0.4.1",
    "pre-commit>=3.6",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = []

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
